Avoid a dodgy unsafeCoerce# with GHC 9.4+
Previously, `isNil` used `unsafeCoerce#` to coerce both of its arguments to
`Any` before calling `reallyUnsafePtrEquality#` on it. This was necessary in
previous GHCs because `reallyUnsafePtrEquality#` had an overly restrictive type
that (1) requiring both of its arguments to have the same type, and (2)
required the type of the arguments to be lifted. This was dangerous, however,
because one of the coerced arguments was of type `SmallMutableArray#`, which
has an unlifted return kind. Coercing to `Any` is only guaranteed to work when
the type being coerced is a lifted type, and `SmallMutableArray#` is not a
lifted type. By a lucky coincidence, this happened to work on previous GHCs,
but this hack finally broke in GHC 9.6, as observed in
https://gitlab.haskell.org/ghc/ghc/-/issues/22813.

Luckily, we no longer need to make use of this hack at all on recent GHCs. This
is because `reallyUnsafePtrEquality#` now has a more general type that is
levity polymorphic and heterogeneous, so we can call `reallyUnsafePtrEquality#`
on its arguments without any intermediate coercions to `Any`. I have applied
the appropriate CPP to make this happen, and I have confirmed that the
`structs` test suite now passes with GHC 9.6.